requirements.txt
psycopg2-binary  
django-heroku  
gunicorn  
django-storages 
boto3 
pathlib
dj_database_url
google-auth
google-cloud-storage
python-dateutil==2.8.0
django-import-export
django-river
djangorestframework
django-rest-auth

